+++ CHANGELOG.md
@@ -15,8 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Graph widgets: Allows scaling the nodes and edges
 - Graph widgets: Allows setting the edge with
 - Graph widgets: Allows exporting nodes to file
-- New model plugin: population growth
-- New model plugin: Conway's Game of Life cellular automaton
+- New model plugin: Population growth `populationGrowth`
+- New model plugin: Conway's Game of Life cellular automaton `gameOfLife`
+- New model plugin: Elementary cellular automaton rules `cellularAutomata1D`
 
 ### Changed
 - Node's coordinates are now a float